The Great SLD Release Date Patch
Using release date information found in taw's decks (which is sourced by
official product page) we can successfully map every single drop.
Bundles and superdrops are not touched but could be derived pretty
easily (like it was done for the Astrology Lands).

Note that this value reflects the release date to the public, aka when
the product actually ships, not when the sale starts or the announcement
was made.